requirements.txt
Django==6.0
django-allauth==65.13.1
django-htmx==1.27.0
django-import-export==4.3.14
django-template-partials==25.3
django-widget-tweaks==1.5.0
pytest==9.0.2
python-dotenv==1.2.1
pandas==2.3.3
django-template-partials==25.3
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
iniconfig==2.3.0
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.4
tablib==3.9.0
psycopg[binary]
yfinance==0.2.66